files/lambda-artifacts/securityhub-suppressor/requirements.txt
aws-lambda-powertools
pylint-gitlab
pyyaml
jmespath
yamllint
